requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.7.4
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.25
langchain-openai==0.0.6
langsmith==0.1.5
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
playsound==1.3.0
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
python-dateutil==2.8.2
python-telegram==0.18.0
python-telegram-bot==20.8
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
telegram_text==0.2.0
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
